+++ CHANGELOG.md
@@ -20,6 +20,15 @@
 - Fix Upload Issue
 - upload any file type
 
+## 0.0.7
+- ADD MDViewer
+
+## 0.0.8
+- ADD MDViewer SaveFile
+
+## 0.0.9
+- ADD MDViewer SaveFile fix issue
+